backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
Pillow==10.4.0
boto3==1.35.36
django-storages==1.14.4